+++ frontend/src/config/routes.ts
@@ -1,0 +1,7 @@
+export default {
+    home: "/",
+    auth: {
+        login: "/login",
+        logout: "/logout",
+    }
+};