+++ frontend/src/pages/Login/LoginPage.tsx
@@ -1,24 +1,33 @@
 import { useState } from 'react';
-import { loginUser } from '../api/auth';
-import type { LoginRequest } from '../types/auth';
+import { useNavigate } from 'react-router-dom';
+import ROUTES from '../../config/routes';
+import { loginUser } from '../../api/auth';
+import type { LoginRequest } from '../../types/auth';
 
 export default function LoginPage() {
     const [credentials, setCredentials] = useState<LoginRequest>({ username: '', password: '' });
     const [error, setError] = useState<string | null>(null);
+    const [loading, setLoading] = useState<boolean>(false);
+    
+    const navigate = useNavigate();
     const handleLogin = async (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault();
         setError(null);
+        setLoading(true);
 
         try {
             await loginUser(credentials);
+            navigate(ROUTES.home);
         } catch (err: any) {
             setError(err.response?.data?.message || 'Login failed');
+        } finally {
+            setLoading(false);
         }
     };
 
     const handleValueChange = (e: React.ChangeEvent<HTMLInputElement>) => {
         const { name, value } = e.target;
-        setCredentials((prev: any) => ({ ...prev, [name]: value }));
+        setCredentials((prev: LoginRequest) => ({ ...prev, [name]: value }));
     };
 
     return (
@@ -35,7 +44,6 @@ export default function LoginPage() {
                     placeholder="username"
                     required
                 />
-                <br />
                 <input
                     type="password"
                     name="password"
@@ -44,8 +52,9 @@ export default function LoginPage() {
                     placeholder="password"
                     required
                 />
-                <br />
-                <button type="submit">Login</button>
+                <button type="submit" disabled={loading}>
+                    {loading ? 'Logging in...' : 'Login'}
+                </button>
             </form>
         </>
     );